updateInfo.js fully refactored
