+++ backend/routes/messaging.js
@@ -162,3 +162,5 @@ router.post("/find-matches", async (req, res) => {
     return res.status(500).json({ error: error.message });
   }
 });
+
+export default router;
