+++ backend/server.js
@@ -8,6 +8,8 @@ import internalOps from "./routes/internalOps.js";
 import signUp from "./routes/signUp.js";
 import login from "./routes/login.js";
 import messaging from "./routes/messaging.js";
+import updateInfo from "./routes/updateInfo.js";
+import getInfo from "./routes/getInfo.js";
 
 const PORT = process.env.PORT || 3000;
 const app = express();
@@ -18,366 +20,8 @@ app.use(internalOps);
 app.use(login);
 app.use(signUp);
 app.use(messaging);
-
-app.post("/get-user-info", async (req, res) => {
-  const token = req.body.accessToken;
-
-  if (!token) {
-    return res.status(403).json({ error: "Access token is required" });
-  }
-
-  try {
-    const { data: user, error } = await supabase.auth.api.getUser(token);
-
-    if (error) {
-      throw error;
-    }
-
-    const { data: userData, fetchError } = await supabase
-      .from("users")
-      .select("*")
-      .eq("uuid", user.id)
-      .single();
-
-    if (fetchError) {
-      throw fetchError;
-    }
-
-    return res.status(200).json({ user: userData });
-  } catch (error) {
-    console.error("Error fetching user data:", error.message);
-    return res.status(401).json({ error: "Unauthorized" });
-  }
-});
-
-app.post("/update-user-info", async (req, res) => {
-  const { accessToken, firstName, lastName, dogName, address, email } =
-    req.body;
-
-  if (!accessToken) {
-    return res.status(401).json({ error: "Access token is required" });
-  }
-
-  try {
-    const { data: user, error } = await supabase.auth.api.getUser(accessToken);
-
-    if (error) {
-      throw error;
-    }
-
-    const { error: updateError } = await supabase
-      .from("users")
-      .update({
-        human_first_name: firstName,
-        human_last_name: lastName,
-        dog_name: dogName,
-        address: address,
-        email: email,
-      })
-      .eq("uuid", user.id);
-
-    if (updateError) {
-      console.error("Update Error:", updateError.message);
-      return res.status(500).json({ error: updateError.message });
-    }
-
-    return res
-      .status(200)
-      .json({ message: "User information updated successfully" });
-  } catch (error) {
-    console.error("Error updating user information:", error.message);
-    return res.status(500).json({ error: error.message });
-  }
-});
-
-app.post("/update-dog-traits", async (req, res) => {
-  const {
-    accessToken,
-    likeability,
-    energy,
-    playfulness,
-    aggression,
-    size,
-    training,
-  } = req.body;
-
-  if (!accessToken) {
-    return res.status(401).json({ error: "Access token is required" });
-  }
-
-  try {
-    const { data: user, error } = await supabase.auth.api.getUser(accessToken);
-    if (error) throw error;
-
-    const { error: updateError } = await supabase
-      .from("userdata")
-      .update({
-        likeability,
-        energy,
-        playfulness,
-        aggression,
-        size,
-        training,
-      })
-      .eq("uuid", user.id);
-
-    if (updateError) {
-      console.error("Update Error:", updateError.message);
-      return res.status(500).json({ error: updateError.message });
-    }
-
-    return res.status(200).json({ message: "Dog traits updated successfully" });
-  } catch (error) {
-    console.error("Error updating dog traits:", error.message);
-    return res.status(500).json({ error: error.message });
-  }
-});
-
-app.post("/update-dog-filter", async (req, res) => {
-  const {
-    accessToken,
-    likeabilityFilter, // Expected to be { min: number, max: number }
-    energyFilter,
-    playfulnessFilter,
-    aggressionFilter,
-    sizeFilter,
-    trainingFilter,
-  } = req.body;
-
-  if (!accessToken) {
-    return res.status(401).json({ error: "Access token is required" });
-  }
-
-  try {
-    const { data: user, error } = await supabase.auth.api.getUser(accessToken);
-    if (error) throw error;
-
-    const { error: updateError } = await supabase
-      .from("userdata")
-      .update({
-        likeabilityFilter,
-        energyFilter,
-        playfulnessFilter,
-        aggressionFilter,
-        sizeFilter,
-        trainingFilter,
-      })
-      .eq("uuid", user.id);
-
-    if (updateError) {
-      console.error("Update Error:", updateError.message);
-      return res.status(500).json({ error: updateError.message });
-    }
-
-    return res
-      .status(200)
-      .json({ message: "Dog traits updated successfully." });
-  } catch (error) {
-    console.error("Error updating dog traits:", error.message);
-    return res.status(500).json({ error: error.message });
-  }
-});
-
-app.post("/current-dog-filters", async (req, res) => {
-  const { accessToken } = req.body;
-
-  if (!accessToken) {
-    return res.status(401).json({ error: "Access token is required" });
-  }
-
-  try {
-    const { data: user, error: userError } = await supabase.auth.api.getUser(
-      accessToken
-    );
-    if (userError) throw userError;
-
-    const { data: userData, error: dataError } = await supabase
-      .from("userdata")
-      .select(
-        "likeabilityFilter, energyFilter, playfulnessFilter, aggressionFilter, sizeFilter, trainingFilter"
-      )
-      .eq("uuid", user.id)
-      .single();
-
-    if (dataError) {
-      throw dataError;
-    }
-
-    return res.status(200).json({ userFilters: userData });
-  } catch (error) {
-    console.error("Error fetching dog traits:", error.message);
-    return res
-      .status(500)
-      .json({ error: "Failed to fetch dog traits: " + error.message });
-  }
-});
-
-app.post("/update-max-distance", async (req, res) => {
-  const { accessToken, maxDistance } = req.body;
-
-  if (!accessToken) {
-    return res.status(401).json({ error: "Access token is required" });
-  }
-
-  try {
-    const { data: user, error } = await supabase.auth.api.getUser(accessToken);
-    if (error) {
-      throw error;
-    }
-
-    const { error: updateError } = await supabase
-      .from("userdata")
-      .update({ maxDistance })
-      .eq("uuid", user.id);
-
-    if (updateError) {
-      console.error("Update Error:", updateError.message);
-      return res.status(500).json({ error: updateError.message });
-    }
-
-    return res
-      .status(200)
-      .json({ message: "Max distance updated successfully" });
-  } catch (error) {
-    console.error("Error updating max distance:", error.message);
-    return res.status(500).json({ error: error.message });
-  }
-});
-
-app.post("/max-distance", async (req, res) => {
-  const { accessToken } = req.body;
-
-  if (!accessToken) {
-    return res.status(401).json({ error: "Access token is required" });
-  }
-
-  try {
-    const { data: user, error: userError } = await supabase.auth.api.getUser(
-      accessToken
-    );
-    if (userError) throw userError;
-
-    const { data, error: dataError } = await supabase
-      .from("userdata")
-      .select("maxDistance")
-      .eq("uuid", user.id)
-      .single();
-
-    if (dataError) {
-      throw dataError;
-    }
-
-    if (data) {
-      return res.status(200).json({ maxDistance: data.maxDistance });
-    } else {
-      return res.status(404).json({ error: "User data not found." });
-    }
-  } catch (error) {
-    console.error("Error fetching max distance:", error.message);
-    return res.status(500).json({ error: error.message });
-  }
-});
-
-app.post("/current-dog-pictures", async (req, res) => {
-  const { accessToken } = req.body;
-
-  if (!accessToken) {
-    return res.status(401).json({ error: "Access token is required" });
-  }
-
-  try {
-    const { data: user, error: userError } = await supabase.auth.api.getUser(
-      accessToken
-    );
-    if (userError) {
-      console.error("Authentication error:", userError.message);
-      return res
-        .status(401)
-        .json({ error: "Authentication failed: " + userError.message });
-    }
-
-    const { data: userData, error: dataError } = await supabase
-      .from("images")
-      .select("*")
-      .eq("uuid", user.id);
-
-    if (dataError) {
-      console.error("Data fetch error:", dataError.message);
-      throw dataError;
-    }
-    console.log(userData);
-    if (userData.length === 0) {
-      return res.status(404).json({ error: "No images found for the user." });
-    }
-
-    return res.status(200).json({ images: userData });
-  } catch (error) {
-    console.error("Error fetching dog images:", error.message);
-    return res
-      .status(500)
-      .json({ error: "Failed to fetch dog images: " + error.message });
-  }
-});
-app.post("/update-dog-pictures", async (req, res) => {
-  const { accessToken, ...pictures } = req.body;
-
-  if (!accessToken) {
-    return res.status(401).json({ error: "Access token is required" });
-  }
-
-  try {
-    const { data: user, error } = await supabase.auth.api.getUser(accessToken);
-    if (error) {
-      console.error("Authentication error:", error.message);
-      return res
-        .status(401)
-        .json({ error: "Authentication failed: " + error.message });
-    }
-
-    const updates = Object.keys(pictures).reduce((acc, key) => {
-      if (key.startsWith("picture") && pictures[key]) {
-        acc[key] = pictures[key];
-      }
-      return acc;
-    }, {});
-
-    if (Object.keys(updates).length === 0) {
-      return res.status(400).json({ error: "No pictures provided for update" });
-    }
-
-    updates.uuid = user.id;
-
-    const { error: upsertError } = await supabase
-      .from("images")
-      .upsert(updates, {
-        returning: "minimal",
-      });
-
-    if (upsertError) {
-      console.error("Upsert Error:", upsertError.message);
-      return res.status(500).json({
-        error:
-          "Failed to update or insert dog pictures: " + upsertError.message,
-      });
-    }
-
-    return res
-      .status(200)
-      .json({ message: "Dog pictures updated successfully." });
-  } catch (error) {
-    console.error("Error uploading image:", error);
-    if (error.response) {
-      console.error("Error data:", error.response.data);
-      console.error("Error status:", error.response.status);
-      console.error("Error headers:", error.response.headers);
-    } else if (error.request) {
-      console.error("No response received:", error.request);
-    } else {
-      console.error("Error setting up the request:", error.message);
-    }
-    console.error("Error config:", error.config);
-  }
-});
+app.use(updateInfo);
+app.use(getInfo);
 
 app.post("/next-user-data", async (req, res) => {
   let nextUserUuid = "";
@@ -508,138 +152,6 @@ app.post("/next-user-data", async (req, res) => {
     distance: distance,
     oou: outofuserstate, //out of users
   });
-});
-
-app.post("/update-bio", async (req, res) => {
-  const { accessToken, bio } = req.body;
-
-  if (!accessToken) {
-    return res.status(401).json({ error: "Access token is required" });
-  }
-
-  if (bio === undefined) {
-    return res.status(400).json({ error: "Bio content is required" });
-  }
-
-  try {
-    const { data: user, error: userError } = await supabase.auth.api.getUser(
-      accessToken
-    );
-    if (userError) throw userError;
-
-    const { error: updateError } = await supabase
-      .from("userdata")
-      .update({ bio: bio })
-      .eq("uuid", user.id);
-
-    if (updateError) {
-      console.error("Update Error:", updateError.message);
-      return res.status(500).json({ error: updateError.message });
-    }
-
-    return res.status(200).json({ message: "Bio updated successfully" });
-  } catch (error) {
-    console.error("Error updating bio:", error.message);
-    return res.status(500).json({ error: error.message });
-  }
-});
-
-app.post("/get-bio", async (req, res) => {
-  const { accessToken } = req.body;
-
-  if (!accessToken) {
-    return res.status(401).json({ error: "Access token is required" });
-  }
-
-  try {
-    const { data: user, error: userError } = await supabase.auth.api.getUser(
-      accessToken
-    );
-    if (userError) throw userError;
-
-    const { data: userData, error: dataError } = await supabase
-      .from("userdata")
-      .select("bio")
-      .eq("uuid", user.id)
-      .single();
-
-    if (dataError) {
-      throw dataError;
-    }
-
-    if (userData) {
-      return res.status(200).json({ bio: userData.bio });
-    } else {
-      return res.status(404).json({ error: "User bio not found." });
-    }
-  } catch (error) {
-    console.error("Error fetching bio:", error.message);
-    return res.status(500).json({ error: error.message });
-  }
-});
-
-app.post("/get-location", async (req, res) => {
-  const { accessToken } = req.body;
-
-  if (!accessToken) {
-    return res.status(401).json({ error: "Access token is required" });
-  }
-
-  try {
-    const { data: user, error: userError } = await supabase.auth.api.getUser(
-      accessToken
-    );
-    if (userError) throw userError;
-
-    const { data: basicInfo, error: basicInfoError } = await supabase
-      .from("users")
-      .select("address")
-      .eq("uuid", user.id)
-      .single();
-
-    if (basicInfoError) {
-      throw basicInfoError;
-    }
-
-    if (!basicInfo.address) {
-      return res.status(404).json({ error: "No address found for this user." });
-    }
-
-    const positionStackApiKey = "be2efb6b90f3a1015d928b4186ca5ec4";
-    const formattedAddress = encodeURIComponent(basicInfo.address);
-    const positionStackUrl = `http://api.positionstack.com/v1/forward?access_key=${positionStackApiKey}&query=${formattedAddress}`;
-
-    const response = await axios.get(positionStackUrl);
-
-    if (!response.data.data || response.data.data.length === 0) {
-      return res
-        .status(404)
-        .json({ error: "No location data found for this address." });
-    }
-
-    const locationData = response.data.data[0];
-
-    const { error: updateError } = await supabase
-      .from("userdata")
-      .update({
-        longitude: locationData.longitude,
-        latitude: locationData.latitude,
-      })
-      .eq("uuid", user.id);
-
-    if (updateError) {
-      throw updateError;
-    }
-
-    return res.status(200).json({
-      message: "Location updated successfully",
-      longitude: locationData.longitude,
-      latitude: locationData.latitude,
-    });
-  } catch (error) {
-    console.error("Error fetching or updating location data:", error.message);
-    return res.status(500).json({ error: error.message });
-  }
 });
 
 app.post("/mark-user-seen", async (req, res) => {
